[JENKINS-8276] Moved method that tests for job filter extenstions from SectionedViewSection to SectionedViewSectionDescriptor.
The config.jelly file was previously attempting to invoke this method in SectionedViewSection via the 'instance'
reference.  However the 'instance' does not exist until the new section is saved.  Because of this the
'Add Job Filter' button could never appear until after the view was saved and re-opened for editing.

By moving the method to the SectionedViewSectionDescriptor class it is available before the section is created by the first save.